Fix bug in container/grant iterator
The specific bug was that the iterator for a container was checking for
null-ness of the pointer to the container rather than the container
itself. As a result, if a container was empty, rather than skipping over
it, it would just dereference 0x0 a cast it to the container type. Bad
news, but this code path hadn't really been tested.

The fix is a bit more expansive. Instead of returning a
pointer-to-a-pointer from `container_for`, we just return a single
pointer so this kind of confusion is harder to make.